fix(insights): expose insightsSummary on analytics to avoid codegen
Generated convex/_generated/api.d.ts is committed and blocked from
edits here. Keep the helper in lib and the public query on the
existing analytics module so api.analytics.insightsSummary typechecks.
