+++ apps/web/app/(tempo)/insights/page.tsx
@@ -3,7 +3,7 @@
  * @category: You
  * @source: docs/design/claude-export/design-system/screens-5.jsx
  * @summary: Personal analytics & trends — read-only insights over tasks, habits, and goals.
- * @queries: insights.summary
+ * @queries: analytics.insightsSummary
  * @mutations: (none)
  * @auth: required (gentle sign-in card otherwise)
  */
